Cleanup folder cache
