+++ FlowCrypt/Functionality/Services/Folders Services/Models/FolderObject.swift
