+++ FlowCrypt/Functionality/Services/Folders Services/Models/FolderViewModel.swift
